+++ src/app/clusters/power-topology-server/power-topology-server.h
@@ -19,52 +19,21 @@
 
 #include <lib/core/Optional.h>
 
-#include <app-common/zap-generated/cluster-objects.h>
+#include "PowerTopologyCluster.h"
 #include <app/AttributeAccessInterface.h>
-#include <lib/support/CommonIterator.h>
 
 namespace chip {
 namespace app {
 namespace Clusters {
 namespace PowerTopology {
 
-class Delegate
-{
-public:
-    virtual ~Delegate() = default;
-
-    /**
-     * Get the Nth EndpointId in the list of available endpoints.
-     * @param index The index of the EndpointId to be returned.
-     * @param endpointId A reference to the EndpointId.
-     * @return Returns a CHIP_NO_ERROR if there was no error and the EndpointId was returned successfully.
-     * CHIP_ERROR_PROVIDER_LIST_EXHAUSTED if the index is greater than or equal to the length of the list of available endpoints.
-     */
-    virtual CHIP_ERROR GetAvailableEndpointAtIndex(size_t index, EndpointId & endpointId) = 0;
-
-    /**
-     * Get the Nth EndpointId in the list of active endpoints.
-     * @param index The index of the EndpointId to be returned.
-     * @param endpointId A reference to the EndpointId.
-     * @return Returns a CHIP_NO_ERROR if there was no error and the EndpointId was returned successfully.
-     * CHIP_ERROR_PROVIDER_LIST_EXHAUSTED if the index is greater than or equal to the length of the list of active endpoints.
-     */
-    virtual CHIP_ERROR GetActiveEndpointAtIndex(size_t index, EndpointId & endpointId) = 0;
-};
-
-enum class OptionalAttributes : uint32_t
-{
-    kOptionalAttributeAvailableEndpoints = 0x1,
-    kOptionalAttributeActiveEndpoints    = 0x2,
-};
-
 class Instance : public AttributeAccessInterface
 {
 public:
     Instance(EndpointId aEndpointId, Delegate & aDelegate, BitMask<Feature> aFeature,
              BitMask<OptionalAttributes> aOptionalAttributes) :
-        AttributeAccessInterface(MakeOptional(aEndpointId), Id),
-        mDelegate(aDelegate), mFeature(aFeature), mOptionalAttrs(aOptionalAttributes)
+        AttributeAccessInterface(MakeOptional(aEndpointId), Id), mDelegate(aDelegate), mFeature(aFeature),
+        mOptionalAttrs(aOptionalAttributes)
     {}
     ~Instance() { Shutdown(); }
